Fix crash from Postgres aggregates arriving as strings
The Neon driver returns NUMERIC and bigint as JavaScript strings, so the
shop list and detail endpoints were handing StarRating a string where the
Shop type promised a number. JSON is not type-checked at runtime, so it
reached value.toFixed(1) and threw "value.toFixed is not a function" as
soon as the seeded reviews gave any shop a rating to display.

The same root cause was hiding a second, silent bug: the review count was
gated on typeof count === 'number', which a string never satisfies, so
the "(4)" beside the stars had never rendered.

Cast in SQL so the response matches the declared type:

  ROUND(AVG(rating)::numeric, 1)::float8 AS rating_avg
  COUNT(*)::int                          AS rating_count

Also coerce inside StarRating. The cast fixes the contract, but the
component consumes untyped JSON and should not crash the page over one
unexpected value.

Verified against the running API: rating_avg and rating_count now come
back as float and int for all ten shops.

+++ app/api/shops/route.ts
@@ -16,7 +16,9 @@ export async function GET(request: NextRequest) {
         COALESCE(rv.rating_count, 0) AS rating_count
       FROM shops s
       LEFT JOIN (
-        SELECT shop_id, ROUND(AVG(rating)::numeric, 1) AS rating_avg, COUNT(*) AS rating_count
+        SELECT shop_id,
+               ROUND(AVG(rating)::numeric, 1)::float8 AS rating_avg,
+               COUNT(*)::int AS rating_count
         FROM reviews
         GROUP BY shop_id
       ) rv ON rv.shop_id = s.id
